Add files via upload
A couple of functions that extend the functionality of the Fraction class in the fractions module.

`frac()` works like `Fraction()` except that instead of returning just the fraction, it returns a tuple of the fraction and
the difference between the fraction and the original value as a proportion of the original value

`simplest_frac()` is similar but will return the simplest fraction (the fraction with the smallest denominator) that is less
than a specified relative difference compared to the original amount. Like `frac()` it returns a tuple containing the
fraction itself and the difference.